requirements.txt
pandas
numpy
scikit-learn
pandas-datareader
matplotlib
yfinance
hmmlearn
fredapiPyYAML
